File-scoped namespaces and layouts cleaning
