+++ TagsCloudApp/ConsoleActions/IUIAction.cs
@@ -1,0 +1,8 @@
+﻿namespace TagsCloudApp.ConsoleActions;
+
+public interface IUIAction
+{
+    string GetDescription();
+
+    void Handle();
+}
