--- TagsCloudApp/ConsoleInterface/ConsoleActions/IUIAction.cs
@@ -1,9 +1,0 @@
-﻿namespace TagsCloudApp.ConsoleInterface.ConsoleActions
-{
-    public interface IUIAction
-    {
-        string GetDescription();
-
-        void Handle();
-    }
-}
