+++ TagsCloudApp/Program.cs
@@ -2,15 +2,14 @@
 using TagsCloudApp.ConsoleInterface;
 using TagsCloudApp.Providers;
 
-namespace TagsCloudApp
+namespace TagsCloudApp;
+
+class Program
 {
-    class Program
+    static void Main(string[] args)
     {
-        static void Main(string[] args)
-        {
-            var container = ContainerProvider.GetContainer();
-            var ui = container.GetInstance<IUI>();
-            ui.Run();
-        }
+        var container = ContainerProvider.GetContainer();
+        var ui = container.GetInstance<IUI>();
+        ui.Run();
     }
 }
