+++ TagsCloudContainer/Infrastructure/CloudSettings.cs
@@ -1,11 +1,10 @@
 ﻿using TagsCloudContainer.Spirals;
 using TagsCloudContainer.TagPainters;
 
-namespace TagsCloudContainer.Infrastructure
+namespace TagsCloudContainer.Infrastructure;
+
+public class CloudSettings
 {
-    public class CloudSettings
-    {
-        public ITagPainter Painter { get; set; }
-        public ISpiral Spiral { get; set; }
-    }
+    public ITagPainter Painter { get; set; }
+    public ISpiral Spiral { get; set; }
 }
