+++ TagsCloudContainer/Spirals/ArchimedeanSpiral.cs
@@ -1,34 +1,33 @@
 ﻿using TagsCloudContainer.Infrastructure;
 
-namespace TagsCloudContainer.Spirals
+namespace TagsCloudContainer.Spirals;
+
+public class ArchimedeanSpiral : ISpiral
 {
-    public class ArchimedeanSpiral : ISpiral
+    public string Name => "Circle spiral";
+    private readonly double spiralCoef;
+    private readonly double angleDelta;
+    private readonly Settings settings;
+    private double currentAngle;
+
+    public ArchimedeanSpiral(Settings settings)
     {
-        public string Name => "Circle spiral";
-        private readonly double spiralCoef;
-        private readonly double angleDelta;
-        private readonly Settings settings;
-        private double currentAngle;
+        this.settings = settings;
+        spiralCoef = 1;
+        angleDelta = Math.PI / 360;
+    }
 
-        public ArchimedeanSpiral(Settings settings)
-        {
-            this.settings = settings;
-            spiralCoef = 1;
-            angleDelta = Math.PI / 360;
-        }
+    public Point GetNext()
+    {
+        var center = settings.Center;
+        var x = Math.Round(spiralCoef * currentAngle * Math.Cos(currentAngle)) + center.X;
+        var y = Math.Round(spiralCoef * currentAngle * Math.Sin(currentAngle)) + center.Y;
+        currentAngle += angleDelta;
+        return new Point((int)x, (int)y);
+    }
 
-        public Point GetNext()
-        {
-            var center = settings.Center;
-            var x = Math.Round(spiralCoef * currentAngle * Math.Cos(currentAngle)) + center.X;
-            var y = Math.Round(spiralCoef * currentAngle * Math.Sin(currentAngle)) + center.Y;
-            currentAngle += angleDelta;
-            return new Point((int)x, (int)y);
-        }
-
-        public void Reset()
-        {
-            currentAngle = 0;
-        }
+    public void Reset()
+    {
+        currentAngle = 0;
     }
 }
