+++ TagsCloudContainerTests/ParserTests/DocParserTests.cs
@@ -1,15 +1,14 @@
 ﻿using NUnit.Framework;
 using TagsCloudContainer.Parsers;
 
-namespace TagsCloudContainerTests.ParserTests
+namespace TagsCloudContainerTests.ParserTests;
+
+internal class DocParserTests : ParserTests
 {
-    internal class DocParserTests : ParserTests
+    [OneTimeSetUp]
+    public void SetUp()
     {
-        [OneTimeSetUp]
-        public void SetUp()
-        {
-            parser = new DocParser();
-            format = "docx";
-        }
+        parser = new DocParser();
+        format = "docx";
     }
 }
