+++ TagsCloudContainerTests/PreprocessorsApplierTests.cs
@@ -6,46 +6,45 @@ using TagsCloudApp.Providers;
 using TagsCloudContainer.Appliers;
 using TagsCloudContainer.Parsers;
 
-namespace TagsCloudContainerTests
+namespace TagsCloudContainerTests;
+
+internal class ApplierTests
 {
-    internal class ApplierTests
+    private string textsFolder;
+    private IParser parser;
+    private IPreprocessorsApplier preprocessorsApplier;
+    private IFiltersApplier filtersApplier;
+
+    [OneTimeSetUp]
+    public void OneTimeSetUp()
     {
-        private string textsFolder;
-        private IParser parser;
-        private IPreprocessorsApplier preprocessorsApplier;
-        private IFiltersApplier filtersApplier;
+        textsFolder = Path.GetFullPath(@"..\..\..\texts");
+        parser = new TxtParser();
+        var settings = SettingsProvider.GetSettings();
+        preprocessorsApplier = new PreprocessorsApplier(settings);
+        filtersApplier = new FiltersApplier(settings);
+    }
 
-        [OneTimeSetUp]
-        public void OneTimeSetUp()
-        {
-            textsFolder = Path.GetFullPath(@"..\..\..\texts");
-            parser = new TxtParser();
-            var settings = SettingsProvider.GetSettings();
-            preprocessorsApplier = new PreprocessorsApplier(settings);
-            filtersApplier = new FiltersApplier(settings);
-        }
+    [Test]
+    public void Should_ApplyPreprocessorsCorrectly()
+    {
+        var path = Path.Combine(textsFolder, "test.txt");
+        var parsed = parser.Parse(path);
+        var preprocessed = preprocessorsApplier.ApplyPreprocessors(parsed.Value);
 
-        [Test]
-        public void Should_ApplyPreprocessorsCorrectly()
-        {
-            var path = Path.Combine(textsFolder, "test.txt");
-            var parsed = parser.Parse(path);
-            var preprocessed = preprocessorsApplier.ApplyPreprocessors(parsed.Value);
+        var result = filtersApplier.ApplyFilters(preprocessed).ToArray();
 
-            var result = filtersApplier.ApplyFilters(preprocessed).ToArray();
+        var expected = new[] {
+            "music",
+            "music",
+            "music",
+            "guitar",
+            "guitar",
+            "piano",
+            "string",
+            "banjo"
+        };
 
-            var expected = new[] {
-                "music",
-                "music",
-                "music",
-                "guitar",
-                "guitar",
-                "piano",
-                "string",
-                "banjo"
-            };
-
-            result.Should().BeEquivalentTo(expected);
-        }
+        result.Should().BeEquivalentTo(expected);
     }
 }
