+++ TagsCloudContainerTests/SpiralTests/OvalSpiralTests.cs
@@ -2,16 +2,15 @@
 using TagsCloudApp.Providers;
 using TagsCloudContainer.Spirals;
 
-namespace TagsCloudContainerTests.SpiralTests
+namespace TagsCloudContainerTests.SpiralTests;
+
+internal class OvalSpiralTests : SpiralTests
 {
-    internal class OvalSpiralTests : SpiralTests
+    [SetUp]
+    public void SetUp()
     {
-        [SetUp]
-        public void SetUp()
-        {
-            var settings = SettingsProvider.GetSettings();
-            center = settings.Center;
-            spiral = new OvalSpiral(settings);
-        }
+        var settings = SettingsProvider.GetSettings();
+        center = settings.Center;
+        spiral = new OvalSpiral(settings);
     }
 }
