+++ TagsCloudContainerTests/TagCloudVisualizationTestsHelper.cs
@@ -2,16 +2,15 @@
 using TagsCloudContainer.Extensions;
 using TagsCloudContainer.TagsCloudLayouter;
 
-namespace TagsCloudContainerTests
+namespace TagsCloudContainerTests;
+
+public static class TagCloudVisualizationTestsHelper
 {
-    public static class TagCloudVisualizationTestsHelper
+    public static double CalculateCloudRadius(this Cloud cloud)
     {
-        public static double CalculateCloudRadius(this Cloud cloud)
-        {
-            return cloud.Rectangles
-                .Select(rectangle => rectangle.Location + rectangle.Size / 2)
-                .Select(rectangleCenter => rectangleCenter.GetDistance(cloud.Center))
-                .Max();
-        }
+        return cloud.Rectangles
+            .Select(rectangle => rectangle.Location + rectangle.Size / 2)
+            .Select(rectangleCenter => rectangleCenter.GetDistance(cloud.Center))
+            .Max();
     }
 }
