+++ TagsCloudContainerTests/TagPainterTests/PainterTests.cs
@@ -6,28 +6,27 @@ using TagsCloudContainer.Infrastructure;
 using TagsCloudContainer.TagPainters;
 using TagsCloudContainer.Tags;
 
-namespace TagsCloudContainerTests.TagPainterTests
+namespace TagsCloudContainerTests.TagPainterTests;
+
+internal abstract class PainterTests
 {
-    internal abstract class PainterTests
+    private readonly Tag[] tags = new[]
     {
-        private readonly Tag[] tags = new[]
-            {
-                new Tag(0.3, "First"),
-                new Tag(0.5, "Second"),
-                new Tag(0.2, "Third")
-            };
-        protected ITagPainter painter;
-        protected Palette palette;
-        protected Func<PaintedTag, int> selector;
-        protected string[] expected;
+        new Tag(0.3, "First"),
+        new Tag(0.5, "Second"),
+        new Tag(0.2, "Third")
+    };
+    protected ITagPainter painter;
+    protected Palette palette;
+    protected Func<PaintedTag, int> selector;
+    protected string[] expected;
 
-        [Test]
-        public void Should_PaintCorrectly()
-        {
-            var result = painter.PaintTags(tags);
+    [Test]
+    public void Should_PaintCorrectly()
+    {
+        var result = painter.PaintTags(tags);
 
-            result.OrderBy(selector).Select(tag => tag.Text)
-                .Should().BeEquivalentTo(expected);
-        }
+        result.OrderBy(selector).Select(tag => tag.Text)
+            .Should().BeEquivalentTo(expected);
     }
 }
